+++ app/src/main/java/lv/yumm/recipes/data/source/LocalRecipe.kt
@@ -16,7 +16,7 @@ data class LocalRecipe (
     val description: String = "",
     val directions: List<String> = emptyList(),
     val complexity: Int = -1,
-    val duration: Int = 0,
+    val duration: Long = 0,
     val portions: Int = 0,
     val imageUrl: String = "",
     val type: RecipeType? = null,
